fix(app_state): scope staleness to the identity and skip rejected sessions
Review findings on #5950.

- The stale flag was keyed on `(api_base, token)` but the age was read from
  a bare global, so identity A succeeding and identity B then failing
  reported A's freshness against B's outage. The ordinary
  `credentials::clear_session` logout also never ran
  `clear_current_user_success` — only the deferred-rejection path did — so
  the leak survived a logout and re-login in one process. The stamp now
  carries the identity it belongs to and is matched on read, which makes it
  self-correcting the way the failure record already was instead of relying
  on every future sign-out path remembering to clear it.

- `DeferredSessionRejected` clears `auth`, `session_token` and
  `current_user`, but the staleness fields were read before that arm ran, so
  a signed-out snapshot could still go out describing a user it no longer
  reports. Excluded at the read.

- A suppressed replay can be wrapping a `TransientResponse` — the backend
  answering badly rather than being unreachable — so the log says
  "unavailable or unhealthy".

- `OPENHUMAN_AUTH_FETCH_TIMEOUT_SECS` is documented in `.env.example` with
  its default and its 2..=12 range, next to the other operator timeouts.

`scripts/ci/check-openhuman-rust-layout.mjs` caps files at 750 lines, so the
growth split out two cohesive units — `ops_auth_timeout.rs` (the fetch
timeout and the backoff step derived from it, kept adjacent because that
derivation is the whole guard against re-opening #5624) and
`ops_staleness.rs` — plus `ops_current_user_backoff_tests.rs` for the
failure/replay/staleness tests. `include!`d files cannot carry `//!` inner
docs, so those headers are `//`.
