+++ .env.example
@@ -104,6 +104,13 @@ OPENHUMAN_TEMPERATURE=0.7
 # in-app "Action timeout" setting (Settings -> Agent OS access), which persists
 # to [agent].agent_timeout_secs. Leave unset to let the UI control it.
 # OPENHUMAN_TOOL_TIMEOUT_SECS=
+# [optional] Operator override for how long the desktop app waits on the
+# backend's current-user (auth_get_me) refresh, in seconds (default 5, range
+# 2-12). Raise it if a slow-but-healthy backend keeps falling back to the
+# stored snapshot. The ceiling is deliberate: this budget is spent inside the
+# app_state_snapshot RPC, and the retry backoff is derived from it so a wider
+# timeout cannot re-open the poll treadmill. Leave unset to use the default.
+# OPENHUMAN_AUTH_FETCH_TIMEOUT_SECS=
 # [optional] Operator override for the inference HTTP request timeout, in
 # seconds (default 120, range 1-3600). Raise this when long reasoning/research
 # turns get cut off at the two-minute mark. Leave unset to use the default.
