+++ src/openhuman/desktop/app_state/ops.rs
@@ -1,6 +1,11 @@
 #[cfg(test)]
+#[path = "ops_current_user_backoff_tests.rs"]
+mod current_user_backoff_tests;
+#[cfg(test)]
 #[path = "ops_tests.rs"]
 mod tests;
+include!("ops_auth_timeout.rs");
 include!("ops_part_01.rs");
+include!("ops_staleness.rs");
 include!("ops_part_02.rs");
 include!("ops_part_03.rs");
