+++ src/openhuman/desktop/app_state/ops_part_01.rs
@@ -43,90 +43,6 @@ const CURRENT_USER_REFRESH_TTL: Duration = Duration::from_secs(5);
 // the agent harness runs on — the agent's turns stalled 50-100s between model
 // calls even though inference itself was idle).
 const RUNTIME_SNAPSHOT_TTL: Duration = Duration::from_secs(10);
-/// Wall-clock budget for one `auth_get_me` refresh when nothing overrides it.
-pub const DEFAULT_AUTH_FETCH_TIMEOUT_SECS: u64 = 5;
-/// Smallest accepted override. Anything shorter would time out a healthy
-/// backend on a merely slow link on every poll — the #5624 treadmill inverted.
-pub const MIN_AUTH_FETCH_TIMEOUT_SECS: u64 = 2;
-/// Largest accepted override.
-///
-/// The ceiling exists because this budget is spent *inside* the snapshot RPC,
-/// concurrently with [`RUNTIME_SNAPSHOT_TIMEOUT`], and the frontend gives that
-/// RPC 30s total. `12 + 10 = 22` leaves headroom; a larger value would let an
-/// operator turn a slow backend into a failed snapshot call.
-pub const MAX_AUTH_FETCH_TIMEOUT_SECS: u64 = 12;
-/// Operator override for [`auth_fetch_timeout`]. A missing, non-numeric or
-/// out-of-range value leaves the default in place (and is logged once).
-pub const AUTH_FETCH_TIMEOUT_ENV_VAR: &str = "OPENHUMAN_AUTH_FETCH_TIMEOUT_SECS";
-
-/// Floor under the first backoff step, independent of the fetch timeout.
-///
-/// Must stay above [`CURRENT_USER_REFRESH_TTL`], which governs how soon a poll
-/// asks for a live fetch at all.
-const CURRENT_USER_BACKOFF_BASE_FLOOR: Duration = Duration::from_secs(10);
-
-/// Parse a raw override into a bounded timeout in seconds.
-///
-/// Pure and global-free so the clamp can be tested without touching the process
-/// environment. `None`, unparseable input, and out-of-range values all fall back
-/// to [`DEFAULT_AUTH_FETCH_TIMEOUT_SECS`].
-pub fn parse_auth_fetch_timeout_secs(raw: Option<&str>) -> u64 {
-    auth_fetch_timeout_override(raw).unwrap_or(DEFAULT_AUTH_FETCH_TIMEOUT_SECS)
-}
-
-/// The override if `raw` names one inside the accepted range, else `None`.
-///
-/// Split from [`parse_auth_fetch_timeout_secs`] so the resolver can tell
-/// "accepted" from "fell back" without re-deriving it by comparing strings —
-/// which would report a valid `05` as ignored.
-fn auth_fetch_timeout_override(raw: Option<&str>) -> Option<u64> {
-    raw.map(str::trim)
-        .and_then(|s| s.parse::<u64>().ok())
-        .filter(|n| (MIN_AUTH_FETCH_TIMEOUT_SECS..=MAX_AUTH_FETCH_TIMEOUT_SECS).contains(n))
-}
-
-/// The effective `auth_get_me` timeout, resolved once per process.
-///
-/// Read through a function rather than held as a `const` so #5930's "5s may be
-/// too tight" has an answer that does not need a rebuild. Every caller — the
-/// `tokio::time::timeout` wrapper, the timeout log line, and the recorded
-/// timeout error's message — reads the same value, so they cannot drift.
-fn auth_fetch_timeout() -> Duration {
-    static RESOLVED: Lazy<Duration> = Lazy::new(|| {
-        let raw = std::env::var(AUTH_FETCH_TIMEOUT_ENV_VAR).ok();
-        match (raw.as_deref(), auth_fetch_timeout_override(raw.as_deref())) {
-            (Some(raw), None) => warn!(
-                "{LOG_PREFIX} ignoring {AUTH_FETCH_TIMEOUT_ENV_VAR}={raw:?}: not an integer in \
-                 {MIN_AUTH_FETCH_TIMEOUT_SECS}..={MAX_AUTH_FETCH_TIMEOUT_SECS}; \
-                 using {DEFAULT_AUTH_FETCH_TIMEOUT_SECS}s"
-            ),
-            (Some(_), Some(secs)) => debug!(
-                "{LOG_PREFIX} auth fetch timeout overridden to {secs}s by {AUTH_FETCH_TIMEOUT_ENV_VAR}"
-            ),
-            (None, _) => {}
-        }
-        Duration::from_secs(parse_auth_fetch_timeout_secs(raw.as_deref()))
-    });
-    *RESOLVED
-}
-
-/// First backoff step after the backend fails to answer `auth_get_me`.
-///
-/// Derived from `fetch_timeout` rather than fixed, because the property that
-/// actually stops the treadmill is *relational*: the step must outlast both the
-/// fetch timeout and the frontend's ~5s `app_state_snapshot` poll, or the next
-/// poll finds the window already expired and pays the full timeout again
-/// (#5624 — 51 timeouts in one session, ~5s each). Making the timeout
-/// configurable (#5930) without deriving this would let an operator re-open
-/// that bug by widening the timeout past a fixed 10s step.
-fn current_user_backoff_base_for(fetch_timeout: Duration) -> Duration {
-    CURRENT_USER_BACKOFF_BASE_FLOOR.max(fetch_timeout.saturating_mul(2))
-}
-
-/// [`current_user_backoff_base_for`] applied to the effective timeout.
-fn current_user_backoff_base() -> Duration {
-    current_user_backoff_base_for(auth_fetch_timeout())
-}
 /// Ceiling on that backoff. Modest on purpose: this window is time during which
 /// a recovered backend still will not be noticed, so it trades a bounded amount
 /// of staleness for not stalling every poll. At the cap a 5s poll loop attempts
@@ -148,15 +64,6 @@ static CURRENT_USER_CACHE: Lazy<Mutex<Option<CachedCurrentUser>>> = Lazy::new(||
 /// identity throughout an outage, and it reads only the positive cache.
 static CURRENT_USER_FAILURE: Lazy<Mutex<Option<CurrentUserFailure>>> =
     Lazy::new(|| Mutex::new(None));
-/// When the backend last answered `auth_get_me` successfully in this process.
-///
-/// Neither existing cache can answer "how old is the user data we are showing":
-/// [`CURRENT_USER_FAILURE`] only knows about failures, and
-/// [`CURRENT_USER_CACHE`] is set to `None` whenever the backend returns an
-/// empty user, which erases the timestamp on a path that is not an outage.
-/// `None` here means "no success yet this process" — the snapshot then reports
-/// staleness with an unknown age rather than inventing one (#5930).
-static LAST_CURRENT_USER_SUCCESS: Lazy<Mutex<Option<Instant>>> = Lazy::new(|| Mutex::new(None));
 static RUNTIME_SNAPSHOT_CACHE: Lazy<Mutex<Option<CachedRuntimeSnapshot>>> =
     Lazy::new(|| Mutex::new(None));
 /// Single-flight gate for the runtime-snapshot rebuild. Concurrent callers whose
@@ -343,41 +250,6 @@ fn record_current_user_failure(api_base: &str, token: &str, error: CurrentUserFe
 /// the stored snapshot after the backend has already come back.
 fn clear_current_user_failure() {
     *CURRENT_USER_FAILURE.lock() = None;
-}
-
-/// Stamp a successful backend answer, so the snapshot can report how old the
-/// data it is serving has become.
-fn note_current_user_success() {
-    *LAST_CURRENT_USER_SUCCESS.lock() = Some(Instant::now());
-}
-
-/// Forget the success stamp on sign-out, so the next account does not inherit
-/// this one's freshness.
-fn clear_current_user_success() {
-    *LAST_CURRENT_USER_SUCCESS.lock() = None;
-}
-
-/// Whether the snapshot is being served from data the backend could not
-/// refresh, and how long since it last did.
-///
-/// Keyed like both caches: a recorded failure only makes *this* `(api_base,
-/// token)` stale. Without the key check, switching environment or signing in as
-/// someone else would inherit the previous identity's outage and report the
-/// fresh data it is about to fetch as stale.
-///
-/// The age is `None` when the backend has not answered at all in this process —
-/// the stored snapshot then came off disk, and its true age is not knowable
-/// from here.
-fn current_user_staleness(api_base: &str, token: &str) -> (bool, Option<u64>) {
-    let stale = CURRENT_USER_FAILURE
-        .lock()
-        .as_ref()
-        .is_some_and(|entry| entry.api_base == api_base && entry.token == token);
-    let age = LAST_CURRENT_USER_SUCCESS
-        .lock()
-        .as_ref()
-        .map(|at| at.elapsed().as_secs());
-    (stale, age)
 }
 
 /// Record the timeout path's failure.
